+++ resources/views/roles/index.blade.php
@@ -4,9 +4,11 @@
             <h2 class="font-semibold text-xl text-gray-800 leading-tight">
                 {{ __('Roles | List') }}
             </h2>
-            <a href="{{ route('roles.create') }}"
-                class="ml-auto btn bg-slate-700 hover:bg-slate-600 text-sm text-white rounded-md px-3 py-2">Create
-                Role</a>
+            @can('create roles')
+                <a href="{{ route('roles.create') }}"
+                    class="ml-auto btn bg-slate-700 hover:bg-slate-600 text-sm text-white rounded-md px-3 py-2">Create
+                    Role</a>
+            @endcan
         </div>
     </x-slot>
 
@@ -46,12 +48,20 @@
                                         <td class="px-5 py-3 text-left">
                                             {{ \Carbon\Carbon::parse($role->created_at)->format('d M Y') }}</td>
                                         <td class="px-5 py-3 text-center">
-                                            <a href="{{ route('roles.edit', $role->id) }}"
-                                                class="bg-slate-700 text-sm text-white rounded-md px-3 py-2 hover:bg-slate-600 mr-2">
-                                                Edit
-                                            </a>
-                                            <a href="javascript:void(0)" onclick="deleteRole({{ $role->id }})"
-                                                class="bg-red-600 text-white text-sm rounded-md px-3 py-2 hover:bg-red-500">Delete</a>
+                                            @canany(['edit roles', 'delete roles'])
+                                                @can('edit roles')
+                                                    <a href="{{ route('roles.edit', $role->id) }}"
+                                                        class="bg-slate-700 text-sm text-white rounded-md px-3 py-2 hover:bg-slate-600 mr-2">
+                                                        Edit
+                                                    </a>
+                                                @endcan
+                                                @can('delete roles')
+                                                    <a href="javascript:void(0)" onclick="deleteRole({{ $role->id }})"
+                                                        class="bg-red-600 text-white text-sm rounded-md px-3 py-2 hover:bg-red-500">Delete</a>
+                                                @endcan
+                                            @else
+                                                ---
+                                            @endcanany
                                         </td>
                                     </tr>
                                 @endforeach
